road-detection-computer-vision/requirements.txt
grpcio
grpcio-tools
opencv-python
msgpack
msgpack_numpy
python-dotenv
deffcode